requirements.txt
fastapi~=0.101.1
transformers
accelerate
diffusers~=0.19.3
xformers==0.0.20
torch==2.0.1
torchvision==0.15.2
numpy~=1.25.2
Pillow~=10.0.0
pydantic~=2.1.1
uvicorn~=0.23.2
python-multipart
matplotlib
opencv-python~=4.8.0.76
boto3~=1.28.21
botocore~=1.31.21
requests~=2.31.0
bitsandbytes
scipy
sentencepiece==0.1.95
protobuf
